refactor the xfr testcases
